eslint: react/no-multi-comp (2)
