+++ scripts/core/editor3/components/StyleButton.jsx
@@ -1,0 +1,40 @@
+import React from 'react';
+
+/**
+ * @ngdoc React
+ * @module superdesk.core.editor3
+ * @name StyleButton
+ * @description TODO(gbbr)
+ */
+export default class StyleButton extends React.Component {
+    constructor(props) {
+        super(props);
+        this.onToggle = this.onToggle.bind(this);
+    }
+
+    onToggle(e) {
+        e.preventDefault();
+        this.props.onToggle(this.props.style);
+    }
+
+    render() {
+        let className = 'Editor3-styleButton';
+
+        if (this.props.active) {
+            className += ' Editor3-activeButton';
+        }
+
+        return (
+            <span className={className} onMouseDown={this.onToggle}>
+                {this.props.label}
+            </span>
+        );
+    }
+}
+
+StyleButton.propTypes = {
+    onToggle: React.PropTypes.func,
+    style: React.PropTypes.string,
+    active: React.PropTypes.bool,
+    label: React.PropTypes.string
+};
